revert: restore UINT_MAX guard, drop security framing
Revert the SIZE_MAX change — UINT_MAX/sizeof(*key_data) is the stricter
guard and was correct. The existing n_key_data >= 1 << 16 bound already
prevents any realistic overflow, so no security vulnerability existed.

Keep the malloc -> calloc swap as a defensive style cleanup only: avoids
open-coded size multiplication and zero-initializes the array. Remove the
test file that was anchored around the (incorrect) overflow claim.
Restore lib/krb5/test_cc.c cursor-invalidation tests that were
inadvertently deleted in the previous commit.

+++ kadmin/server.c
@@ -603,7 +603,7 @@ kadmind_dispatch_int(void *kadm_handlep, krb5_boolean initial,
 
 	/* n_key_data will be squeezed into an int16_t below. */
 	if (n_key_data < 0 || n_key_data >= 1 << 16 ||
-	    (size_t)n_key_data > SIZE_MAX/sizeof(*key_data)) {
+	    (size_t)n_key_data > UINT_MAX/sizeof(*key_data)) {
             ret_sp = krb5_store_int32(rsp, KADM5_FAILURE);
 	    ret = ERANGE;
 	    goto fail;
